+++ CHANGELOG.md
@@ -52,11 +52,25 @@ This is the current release cycle, so stay tuned for future releases!
 - **Add the SQLConnector flavor `postgis`.**  
   The new flavor `postgis` (built atop the `postgresql` flavor) allows pipes to take natively support `GEOMETRY` (and `GEOGRAPHY`) types.
 
+- **Add a pages sidebar to the Web Console.**  
+  Clicking the Meerschaum logo on the Web Console will show the pages navigation sidebar, allowing you to easily expand the web app. Custom pages added via `@web_page` are grouped by plugin. You can override this behavior with the `page_group` parameter:
+
+  ```python
+  from meerschaum.plugins import web_page, dash_plugin
+
+  @dash_plugin
+  def init_dash(dash_app):
+      import dash.html as html
+
+      @web_page('users', page_group='Settings', login_required=True)
+      def my_new_settings_page():
+          return dbc.Container([
+              html.P("A new page `/dash/users/` added to the Settings page group.")
+          ])
+  ```
+
 - **Add the property `instance_keys` to `api` connectors.**  
   The optional property `instance_keys` determines the value of `instance_keys` to be sent alongside pipe requests.
-
-- **Add a pages sidebar to the Web Console.**  
-  Clicking the Meerschaum logo on the Web Console will show the pages navigation sidebar, allowing for future functionality to reside within separate pages. Note that custom pages added via `@web_page` are included as well.
 
 - **Insert the Web Console navbar into custom pages.**  
   Custom pages added via `@web_page` will now include the simple navbar by default, to more tightly integrate custom pages into the Web Console. This behavior may be disabled by setting `skip_navbar=True` in `@web_page`: